Propogate CPM_LIBRARIES upwards.
With this change, target_link_libraries is no longer required in
modules. Just a call to target_link_libraries at the application level
with ${CPM_LIBRARIES} as an argument is all that is required.

This addresses build problems that used to arise when using emscripten's
emmake with CPM. It would think that symbols were multiply defined if
the same library was linked using multiple calls to target_link_library.